requirements.txt
ultralytics
supervision
numpy
opencv-python
PyYAML